pyproject.toml
[project]
name = "life-expectancy-forecast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.66.1",
    "evidently>=0.7.20",
    "fastapi>=0.133.1",
    "google-genai>=1.65.0",
    "joblib>=1.5.3",
    "langchain-google-genai>=4.2.1",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "mlflow>=3.10.0",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "plotly>=5.24.1",
    "pymupdf>=1.27.1",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "redis>=7.2.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "streamlit>=1.54.0",
    "tabula-py>=2.10.0",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.41.0",
    "xgboost>=3.2.0",
]
